Componente de DefaultLayout aplicado a todas as rotas.
