--- src/components/pages/dashboard/index.ts
@@ -1,3 +1,0 @@
-export { IncomeChart } from './Charts/incomeChart'
-export { OutcomeChart } from './Charts/outcomeChart'
-export { CountIncomeOutcomeChart } from './Charts/countIncomeOutcome'
